rename */Float to */Double, for PHP 7 future-compat

+++ src/Rule/Sanitize/Double.php
@@ -15,7 +15,7 @@ namespace Aura\Filter\Rule\Sanitize;
  * @package Aura.Filter
  *
  */
-class Float
+class Double
 {
     /**
      *
